requirements.txt
absl-py==0.8.0
astor==0.8.0
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
en-core-web-sm==2.1.0
gast==0.2.2
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mpmath==1.1.0
murmurhash==1.0.2
networkx==2.3
nltk==3.4.5
nose==1.3.7
numpy==1.17.1
pandas==0.25.1
Pillow==6.2.2
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
pycurl==7.43.0
pygobject==3.20.0
pyparsing==2.4.2
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
spacy==2.1.8
srsly==0.1.0
sympy==1.4
termcolor==1.1.0
thinc==7.0.8
torch==1.1.0
torchvision==0.4.0
tqdm==4.35.0
unattended-upgrades==0.1
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.6
scikit-learn
matplotlib